+++ fjs/fsm/module.f.ts
@@ -4,7 +4,7 @@
  * @module
  */
 import { equal, isEmpty, fold, toArray, scan, foldScan, empty as emptyList, type List } from '../types/list/module.f.ts'
-import type { StringMap } from '../types/object/module.f.ts'
+import type { Map } from '../types/object/module.f.ts'
 import { toRangeMap, union as byteSetUnion, one, empty, range, type ByteSet } from '../types/byte_set/module.f.ts'
 import { intersect, type SortedSet, union as sortedSetUnion } from '../types/sorted_set/module.f.ts'
 import {
@@ -25,7 +25,7 @@ type Rule = readonly [string, ByteSet, string]
 
 export type Grammar = List<Rule>
 
-type Dfa = StringMap<string, RangeMapArray<string>>
+type Dfa = Map<RangeMapArray<string>>
 
 const stringifyIdentity = stringify(identity)